pyproject.toml
[project]
name = "auto-mailing-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Fazliddin Abdurahimov",email = "fazliddin1801@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "xxhash (>=3.5.0,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "auto_mailing_backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"
psycopg2-binary = "^2.9.10"
alembic-postgresql-enum = "^1.7.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev = "scripts.dev:main"
